Decouple `InMemoryModuleMetadataCache` from `DefaultModuleMetadataCache`
Isolate the in-memory caching from the persistent caching, so that we can
later use a different in-memory cache for all projects.
